requirements.txt
python-telegram-bot[job-queue]>=22.0
httpx>=0.27
